interview-to-habr/requirements.txt
google-generativeai>=0.5.0
textual>=0.50.0
rich>=13.0.0
python-docx>=1.0.0
pyyaml>=6.0
click>=8.0.0
pydantic>=2.0.0
